crop_disease_project/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.1.0
numpy>=1.24.3
pandas>=2.0.3
tqdm>=4.66.0

matplotlib>=3.8.0

# For evaluation metrics used in model/train_model.py
scikit-learn>=1.3.0

fastapi
uvicorn[standard]
pydantic>=2.0.0
torch
torchvision
numpy
pillow
scikit-learn
matplotlib
python-multipart
